Add files via upload
Disclaimer: This program is provided for educational and research purposes only. The creator of this program does not condone or support any illegal or malicious activity, and will not be held responsible for any such actions taken by others who may use this program. By downloading or using this program, you acknowledge that you are solely responsible for any consequences that may result from the use of this program.

Note: If the grabber fails to function, attempt rebuilding it without enabling the "Anti VM" option.

Note 2: To decrease the chances of detections, you can download MingGW-w64 (with GCC) and include the "bin" folder in the PATH variable.